fix(model, service) : 과거 예약 점검 시점 변경
- 과거 예약 요청은 불가
- 과거 예약에 대한 조회 및 객체 생성은 가능
