+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,9 +1,12 @@
 package roomescape.service;
 
+import java.time.DateTimeException;
+import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.dto.ReservationRequestDto;
 import roomescape.model.Reservation;
+import roomescape.model.ReservationDate;
 import roomescape.model.ReservationTime;
 import roomescape.repository.ReservationRepository;
 
@@ -25,6 +28,7 @@ public class ReservationService {
     public Reservation addReservation(ReservationRequestDto reservationRequestDto) {
         ReservationTime reservationTime = reservationTimeService.getReservationTimeById(
                 reservationRequestDto.timeId());
+        validateDateTime(ReservationDate.of(reservationRequestDto.date()), reservationTime);
         return reservationRepository.addReservation(reservationRequestDto, reservationTime);
     }
 
@@ -32,4 +36,12 @@ public class ReservationService {
         return reservationRepository.deleteReservation(id);
     }
 
+    private void validateDateTime(ReservationDate reservationDate, ReservationTime reservationTime) {
+
+        LocalDateTime dateTime = LocalDateTime.of(reservationDate.getDate(), reservationTime.getStartAt());
+        LocalDateTime now = LocalDateTime.now();
+        if (dateTime.isBefore(now)) {
+            throw new DateTimeException("과거 예약은 불가능합니다.");
+        }
+    }
 }
